refactor(ledger): drop Huma suffixes on the last non-tx resources :hammer:
Converts the six remaining non-transaction resources in one commit: operation
(3 terminals), metadata (3), encryption (2), audit (1), count (1) and
composition (1). They land together because their registrars share routes.go,
crm_routes.go and composition_routes.go, and their permissions.yaml citations
sit in one file.

Every terminal was verified unreachable first: each registrar attaches the auth
chain as middleware only and hands its terminals to Huma. Encryption and audit
additionally sit behind the envelope-mode nil-handler guard, which is
unchanged — in legacy mode neither the middleware nor the Huma terminals
register.

Coverage holds or improves everywhere. Three cores now exceed their old
baseline because a branch was unreachable from both transports before:
audit.getAuditEvents (94.3% -> 100%, inverted date range),
metadata.getAllMetadataIndexes (87.1% -> 95.2%), and operation's
metadata-filter 404. operation.updateOperation was at 0% on the Huma side and
is now 100%.

Renames follow the established convention: methods drop the suffix, envelope
types become Request/Response, and each transport file takes <resource>_handler.go.

Dead code removed with the terminals: count's parseCountFilter, whose only
caller was the deleted terminal. Its fuzz and property tests were ported to
call buildCountFilter through url.ParseQuery rather than dropped.

transaction_block_unblock_integration_test.go mounted the operation terminal
this commit deletes; it now goes through RegisterOperationRoutes. Fiber v3
takes an untyped any for handlers, so that mount compiled fine and would only have
failed at run time — both affected integration tests were executed against
testcontainers to confirm, not just compiled.

Stale registrar comments in routes.go and crm_routes.go are rewritten for the
converted resources. The transaction and transaction-route entries still say
Huma-migrated and are left for that conversion.

No route, operation ID, or wire contract changes: check-docs reports no
drift.

X-Lerian-Ref: 0x1
