+++ components/ledger/internal/adapters/http/in/composition_handler_test.go
@@ -6,6 +6,7 @@ package in
 
 import (
 	"bytes"
+	"context"
 	"encoding/json"
 	"io"
 	"net/http"
@@ -25,6 +26,26 @@ import (
 	"github.com/LerianStudio/midaz/v4/pkg/mmodel"
 	pkgHTTP "github.com/LerianStudio/midaz/v4/pkg/net/http"
 )
+
+// stubAccountCreator satisfies composition.AccountCreator for handler tests.
+type stubAccountCreator struct {
+	account *mmodel.Account
+	err     error
+}
+
+func (s stubAccountCreator) CreateAccount(_ context.Context, _, _ uuid.UUID, _ *mmodel.CreateAccountInput, _ string) (*mmodel.Account, error) {
+	return s.account, s.err
+}
+
+// stubInstrumentCreator satisfies composition.InstrumentCreator for handler tests.
+type stubInstrumentCreator struct {
+	instrument *mmodel.Instrument
+	err        error
+}
+
+func (s stubInstrumentCreator) CreateInstrument(_ context.Context, _ string, _ uuid.UUID, _ *mmodel.CreateInstrumentInput) (*mmodel.Instrument, error) {
+	return s.instrument, s.err
+}
 
 // buildHumaCompositionApp mounts the single composition Huma operation on a /v2
 // group, faithfully mirroring the production wiring in unified-server.go:
@@ -83,7 +104,7 @@ func validCompositionBody() []byte {
 	return body
 }
 
-func TestHuma_CreateHolderAccount_Success(t *testing.T) {
+func TestCreateHolderAccount_Success(t *testing.T) {
 	// NOT parallel: buildHumaCompositionApp mutates process-global huma state.
 	orgID := uuid.New()
 	ledgerID := uuid.New()
@@ -119,7 +140,7 @@ func TestHuma_CreateHolderAccount_Success(t *testing.T) {
 	assert.Equal(t, createdAccount.ID, acc["id"])
 }
 
-func TestHuma_CreateHolderAccount_AuthPreserved(t *testing.T) {
+func TestCreateHolderAccount_AuthPreserved(t *testing.T) {
 	// NOT parallel: process-global huma state.
 	orgID := uuid.New()
 	ledgerID := uuid.New()
@@ -140,7 +161,7 @@ func TestHuma_CreateHolderAccount_AuthPreserved(t *testing.T) {
 	assert.Equal(t, http.StatusUnauthorized, resp.StatusCode, "auth middleware must reject before Huma; no public route")
 }
 
-func TestHuma_CreateHolderAccount_ValidationError_Canonical400(t *testing.T) {
+func TestCreateHolderAccount_ValidationError_Canonical400(t *testing.T) {
 	// NOT parallel: process-global huma state.
 	orgID := uuid.New()
 	ledgerID := uuid.New()
@@ -171,7 +192,7 @@ func TestHuma_CreateHolderAccount_ValidationError_Canonical400(t *testing.T) {
 	assert.Equal(t, float64(http.StatusBadRequest), got["status"])
 }
 
-func TestHuma_CreateHolderAccount_MalformedBody_Canonical400(t *testing.T) {
+func TestCreateHolderAccount_MalformedBody_Canonical400(t *testing.T) {
 	// NOT parallel: process-global huma state.
 	orgID := uuid.New()
 	ledgerID := uuid.New()
@@ -199,7 +220,7 @@ func TestHuma_CreateHolderAccount_MalformedBody_Canonical400(t *testing.T) {
 	assert.Equal(t, float64(http.StatusBadRequest), got["status"])
 }
 
-func TestHuma_CreateHolderAccount_BadUUID_Canonical400(t *testing.T) {
+func TestCreateHolderAccount_BadUUID_Canonical400(t *testing.T) {
 	// NOT parallel: process-global huma state.
 	orgID := uuid.New()
 	ledgerID := uuid.New()
@@ -227,7 +248,7 @@ func TestHuma_CreateHolderAccount_BadUUID_Canonical400(t *testing.T) {
 	assert.Equal(t, cn.ErrInvalidPathParameter.Error(), got["code"])
 }
 
-func TestHuma_CreateHolderAccount_BusinessError_Preserved(t *testing.T) {
+func TestCreateHolderAccount_BusinessError_Preserved(t *testing.T) {
 	// NOT parallel: process-global huma state. The account-create fails with a
 	// business error; HumaProblem must project the canonical envelope verbatim.
 	orgID := uuid.New()
@@ -257,4 +278,124 @@ func TestHuma_CreateHolderAccount_BusinessError_Preserved(t *testing.T) {
 	var got map[string]any
 	require.NoError(t, json.Unmarshal(respBody, &got), "body: %s", string(respBody))
 	assert.Equal(t, cn.ErrAssetCodeNotFound.Error(), got["code"], "business error code preserved across Huma")
+}
+
+func TestCreateHolderAccount_WithInstrument_201(t *testing.T) {
+	// NOT parallel: process-global huma state.
+	orgID := uuid.New()
+	ledgerID := uuid.New()
+	holderID := uuid.New()
+
+	createdAccount := &mmodel.Account{ID: uuid.New().String(), AssetCode: "USD", Type: "deposit"}
+	instrumentID := uuid.New()
+
+	handler := &CompositionHandler{Service: composition.NewService(
+		stubAccountCreator{account: createdAccount},
+		stubInstrumentCreator{instrument: &mmodel.Instrument{ID: &instrumentID}},
+	)}
+
+	app := buildHumaCompositionApp(t, handler, true)
+
+	body, _ := json.Marshal(map[string]any{
+		"assetCode":      "USD",
+		"type":           "deposit",
+		"bankingDetails": map[string]any{},
+	})
+
+	req := httptest.NewRequest(http.MethodPost, compositionURL(orgID, ledgerID, holderID), bytes.NewReader(body))
+	req.Header.Set("Content-Type", "application/json")
+
+	resp, err := app.Test(req, fiber.TestConfig{Timeout: 0})
+	require.NoError(t, err)
+	defer func() { _ = resp.Body.Close() }()
+
+	respBody, _ := io.ReadAll(resp.Body)
+
+	assert.Equal(t, http.StatusCreated, resp.StatusCode)
+
+	var got mmodel.HolderAccountResponse
+	require.NoError(t, json.Unmarshal(respBody, &got), "body: %s", string(respBody))
+	require.NotNil(t, got.Account)
+	require.NotNil(t, got.Instrument, "instrument created alongside the account")
+	assert.Nil(t, got.InstrumentError)
+}
+
+// TestCreateHolderAccount_PartialFailure_201 locks the partial-failure
+// contract: the account is committed, the instrument write fails, and the service
+// returns a nil error, so the terminal renders 201 carrying the typed failure block.
+func TestCreateHolderAccount_PartialFailure_201(t *testing.T) {
+	// NOT parallel: process-global huma state.
+	orgID := uuid.New()
+	ledgerID := uuid.New()
+	holderID := uuid.New()
+
+	createdAccount := &mmodel.Account{ID: uuid.New().String(), AssetCode: "USD", Type: "deposit"}
+
+	handler := &CompositionHandler{Service: composition.NewService(
+		stubAccountCreator{account: createdAccount},
+		stubInstrumentCreator{err: pkg.ValidateBusinessError(cn.ErrEntityNotFound, "Holder")},
+	)}
+
+	app := buildHumaCompositionApp(t, handler, true)
+
+	body, _ := json.Marshal(map[string]any{
+		"assetCode":      "USD",
+		"type":           "deposit",
+		"bankingDetails": map[string]any{},
+	})
+
+	req := httptest.NewRequest(http.MethodPost, compositionURL(orgID, ledgerID, holderID), bytes.NewReader(body))
+	req.Header.Set("Content-Type", "application/json")
+
+	resp, err := app.Test(req, fiber.TestConfig{Timeout: 0})
+	require.NoError(t, err)
+	defer func() { _ = resp.Body.Close() }()
+
+	respBody, _ := io.ReadAll(resp.Body)
+
+	assert.Equal(t, http.StatusCreated, resp.StatusCode, "a failed instrument does not roll back the committed account")
+
+	var got mmodel.HolderAccountResponse
+	require.NoError(t, json.Unmarshal(respBody, &got), "body: %s", string(respBody))
+	require.NotNil(t, got.Account, "account remains persisted on instrument failure")
+	assert.Nil(t, got.Instrument)
+	require.NotNil(t, got.InstrumentError, "typed failure block surfaced")
+	assert.Equal(t, "FAILED", got.InstrumentError.Status)
+	assert.Equal(t, cn.ErrEntityNotFound.Error(), got.InstrumentError.Reason)
+}
+
+// TestCreateHolderAccount_BadPathUUID_Direct drives the terminal's defensive
+// org/ledger/holder guards, which the wired ParseUUIDPathParameters middleware makes
+// unreachable through the app.
+func TestCreateHolderAccount_BadPathUUID_Direct(t *testing.T) {
+	t.Parallel()
+
+	handler := &CompositionHandler{Service: composition.NewService(stubAccountCreator{}, stubInstrumentCreator{})}
+
+	tests := []struct {
+		name string
+		in   *CreateHolderAccountRequest
+	}{
+		{
+			name: "bad organization_id",
+			in:   &CreateHolderAccountRequest{OrganizationID: "not-a-uuid", LedgerID: uuid.New().String(), ID: uuid.New().String()},
+		},
+		{
+			name: "bad holder id",
+			in:   &CreateHolderAccountRequest{OrganizationID: uuid.New().String(), LedgerID: uuid.New().String(), ID: "not-a-uuid"},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
+
+			_, err := handler.CreateHolderAccount(context.Background(), tt.in)
+
+			var detail *pkgHTTP.Detail
+			require.ErrorAs(t, err, &detail, "terminal must return the canonical problem detail")
+			assert.Equal(t, http.StatusBadRequest, detail.Status)
+			assert.Equal(t, cn.ErrInvalidPathParameter.Error(), detail.Code)
+		})
+	}
 }
