+++ components/ledger/internal/adapters/http/in/count_transactions_by_filters_fuzz_test.go
@@ -5,10 +5,9 @@
 package in
 
 // =============================================================================
-// FUZZ TESTS — parseCountFilter Input Parsing
+// FUZZ TESTS — count filter input parsing
 //
-// These tests exercise the parseCountFilter function with fuzzer-generated
-// inputs to verify:
+// These tests exercise buildCountFilter with fuzzer-generated inputs to verify:
 //   1. No panics on any combination of route, status, start_date, end_date.
 //   2. Validation errors are returned correctly for invalid inputs.
 //   3. No SQL injection vectors pass through validation.
@@ -24,26 +23,15 @@ import (
 	"fmt"
 	"net/url"
 	"strings"
-	"sync"
 	"testing"
 	"time"
 
-	"github.com/gofiber/fiber/v3"
 	"github.com/stretchr/testify/assert"
-	"github.com/valyala/fasthttp"
 )
 
-// fuzzCountApp is the shared Fiber app used to acquire a lightweight ctx per
-// fuzz execution. It is built once (sync.OnceValue, lazily on first fuzz call)
-// so the per-exec cost is just AcquireCtx/ReleaseCtx rather than a full
-// fiber.New() plus an app.Test() HTTP round-trip. That round-trip's per-exec
-// latency is what provoked the Go -fuzztime boundary "context deadline
-// exceeded" flake; keeping each execution cheap removes it at the source.
-var fuzzCountApp = sync.OnceValue(func() *fiber.App { return fiber.New() })
-
-// fuzzParseCountFilter is a test helper that creates a Fiber context with the
-// given query parameters and calls parseCountFilter. It returns the error (if
-// any) and a boolean indicating whether the function panicked.
+// fuzzParseCountFilter encodes the given values into a query string, re-parses it
+// the way the Huma binder does (url.Values), and calls buildCountFilter. It returns
+// the error (if any) and a boolean indicating whether the call panicked.
 func fuzzParseCountFilter(t *testing.T, route, status, startDate, endDate string) (err error, panicked bool) {
 	t.Helper()
 
@@ -71,29 +59,20 @@ func fuzzParseCountFilter(t *testing.T, route, status, startDate, endDate string
 		qv.Set("end_date", endDate)
 	}
 
-	target := "/test"
-	if len(qv) > 0 {
-		target += "?" + qv.Encode()
+	// Round-trip through encode/decode so the fuzzer exercises the exact values a
+	// wire query yields, which is what the Huma binder reads.
+	decoded, parseErr := url.ParseQuery(qv.Encode())
+	if parseErr != nil {
+		return parseErr, false
 	}
 
-	// parseCountFilter reads only query parameters via c.Query, so acquiring a
-	// Fiber ctx with the request URI set exercises the exact same parsing path as
-	// a full app.Test() round-trip — but at orders-of-magnitude higher throughput.
-	app := fuzzCountApp()
-
-	fctx := &fasthttp.RequestCtx{}
-	fctx.Request.SetRequestURI(target)
-
-	c := app.AcquireCtx(fctx)
-	defer app.ReleaseCtx(c)
-
-	_, capturedErr := parseCountFilter(c)
+	_, capturedErr := buildCountFilter(decoded.Get("route"), decoded.Get("status"), decoded.Get("start_date"), decoded.Get("end_date"))
 
 	return capturedErr, false
 }
 
 // FuzzParseCountFilter_Route fuzzes the route query parameter to verify
-// parseCountFilter never panics regardless of input.
+// buildCountFilter never panics regardless of input.
 //
 // Invariants verified:
 //   - No panic on any input
@@ -116,7 +95,7 @@ func FuzzParseCountFilter_Route(f *testing.F) {
 		}
 
 		err, panicked := fuzzParseCountFilter(t, route, "", "", "")
-		assert.False(t, panicked, "parseCountFilter must not panic for route=%q", route)
+		assert.False(t, panicked, "buildCountFilter must not panic for route=%q", route)
 
 		// Route has no validation (any string is accepted).
 		// The only error source is date parsing, and we provide no dates.
@@ -151,9 +130,9 @@ func FuzzParseCountFilter_Status(f *testing.F) {
 		}
 
 		err, panicked := fuzzParseCountFilter(t, "", status, "", "")
-		assert.False(t, panicked, "parseCountFilter must not panic for status=%q", status)
+		assert.False(t, panicked, "buildCountFilter must not panic for status=%q", status)
 
-		// parseCountFilter trims the status before checking.
+		// buildCountFilter trims the status before checking.
 		// If the trimmed result is empty, no filter is applied (valid).
 		trimmed := strings.TrimSpace(status)
 		if trimmed == "" {
@@ -211,9 +190,9 @@ func FuzzParseCountFilter_Dates(f *testing.F) {
 		}
 
 		err, panicked := fuzzParseCountFilter(t, "", "", startDate, endDate)
-		assert.False(t, panicked, "parseCountFilter must not panic for dates start=%q end=%q", startDate, endDate)
+		assert.False(t, panicked, "buildCountFilter must not panic for dates start=%q end=%q", startDate, endDate)
 
-		// parseCountFilter trims whitespace from both dates.
+		// buildCountFilter trims whitespace from both dates.
 		// If the trimmed result is empty, defaults are used (valid).
 		trimmedStart := strings.TrimSpace(startDate)
 		trimmedEnd := strings.TrimSpace(endDate)
@@ -235,7 +214,7 @@ func FuzzParseCountFilter_Dates(f *testing.F) {
 		}
 
 		// Both parseable or empty — enforce start <= end invariant.
-		// parseCountFilter defaults missing dates to today's UTC range,
+		// buildCountFilter defaults missing dates to today's UTC range,
 		// so we must compute the effective dates the same way.
 		now := time.Now().UTC()
 		var effectiveStart, effectiveEnd time.Time
@@ -292,7 +271,7 @@ func FuzzParseCountFilter_AllParams(f *testing.F) {
 		}
 
 		err1, panicked := fuzzParseCountFilter(t, route, status, startDate, endDate)
-		assert.False(t, panicked, "parseCountFilter must not panic")
+		assert.False(t, panicked, "buildCountFilter must not panic")
 
 		// Determinism: same input must produce same result
 		err2, panicked2 := fuzzParseCountFilter(t, route, status, startDate, endDate)
