+++ components/ledger/internal/adapters/http/in/metadata_handler.go
@@ -17,27 +17,25 @@ import (
 	pkgHTTP "github.com/LerianStudio/midaz/v4/pkg/net/http"
 )
 
-// This file is the ledger's Huma adoption of the metadata-index (settings) resource,
+// This file holds the Huma terminals for the metadata-index (settings) resource,
 // following the asset exemplar (asset_handler.go). Metadata differs from asset
 // in two ways: it lives under /v1/settings (no org/ledger path, no UUID path params,
 // so NO ParseUUIDPathParameters middleware), and its path params are plain strings
-// (entity_name enum, index_key). The proven conventions still hold:
+// (entity_name enum, index_key). The shared conventions:
 //
 //  1. Body ops carry RawBody []byte + SkipValidateBody so the imperative
-//     http.DecodeAndValidate (the SAME pipeline the Fiber WithBody decorator runs)
-//     stays the sole body validator — never a native Huma 422.
+//     http.DecodeAndValidate stays the sole body validator — never a native Huma 422.
 //  2. List captures the raw query (via Resolve) and rebuilds the map[string]string
-//     that the shared core feeds to http.ValidateParameters, byte-identical to the
-//     Fiber c.Queries() path.
-//  3. Errors go through the shared pkgHTTP.HumaProblem (RFC 9457 problem+json,
-//     field/status/code-identical to the Fiber http.WithError path).
-//  4. Auth stays a Fiber middleware chain (auth.Authorize("midaz","settings",verb) +
+//     that the shared core feeds to http.ValidateParameters.
+//  3. Errors go through the shared pkgHTTP.HumaProblem (RFC 9457 problem+json), which
+//     fixes each canonical Midaz error at one code and one HTTP status.
+//  4. Auth is a Fiber middleware chain (auth.Authorize("midaz","settings",verb) +
 //     tenant PostAuthMiddlewares) attached in routes.go/unified-server.go BEFORE the
 //     Huma registration — NOT a Huma Security scheme. The per-op Security metadata
 //     below is SPEC-ONLY (for the generated OAS/SDK).
 //
 // The transport-agnostic cores (createMetadataIndex / getAllMetadataIndexes /
-// deleteMetadataIndex) live in metadata.go and are shared with the Fiber wrappers.
+// deleteMetadataIndex) live in metadata.go.
 
 // secMetadataBearerOrAPIKey advertises that each metadata operation accepts EITHER a
 // JWT bearer token OR an X-API-Key (two entries = OR). SPEC metadata only; runtime
@@ -49,26 +47,25 @@ var secMetadataBearerOrAPIKey = []map[string][]string{
 
 // --- POST /settings/metadata-indexes/entities/{entity_name} -------------------
 
-// CreateMetadataIndexInputHuma is the Huma request envelope for POST. RawBody keeps
+// CreateMetadataIndexRequest is the Huma request envelope for POST. RawBody keeps
 // the body out of Huma's validator (see file header); entity_name is a plain string
 // path param (no UUID, so no format tag).
-type CreateMetadataIndexInputHuma struct {
+type CreateMetadataIndexRequest struct {
 	EntityName string `path:"entity_name" doc:"Entity name (organization, ledger, segment, account, portfolio, asset, account_type, transaction, operation, operation_route, transaction_route)"`
 	RawBody    []byte `contentType:"application/json"`
 }
 
-// CreateMetadataIndexOutputHuma pins 201 (matching http.Created).
-type CreateMetadataIndexOutputHuma struct {
+// CreateMetadataIndexResponse pins 201 (matching http.Created).
+type CreateMetadataIndexResponse struct {
 	Status int
 	Body   *mmodel.MetadataIndex
 }
 
-// CreateMetadataIndexHuma decodes+validates the raw body imperatively then delegates
+// CreateMetadataIndex decodes+validates the raw body imperatively then delegates
 // to the shared createMetadataIndex core. It passes an empty query map: the POST
-// route has no meaningful query params (the Fiber path validated c.Queries(), which
-// is empty here), so ValidateParameters over an empty map is a no-op that preserves
-// the canonical flow.
-func (handler *MetadataIndexHandler) CreateMetadataIndexHuma(ctx context.Context, in *CreateMetadataIndexInputHuma) (*CreateMetadataIndexOutputHuma, error) {
+// route carries no query params, so the core's ValidateParameters call is a no-op
+// that keeps the canonical flow identical across the three operations.
+func (handler *MetadataIndexHandler) CreateMetadataIndex(ctx context.Context, in *CreateMetadataIndexRequest) (*CreateMetadataIndexResponse, error) {
 	payload := new(mmodel.CreateMetadataIndexInput)
 	if _, err := pkgHTTP.DecodeAndValidate(in.RawBody, payload); err != nil {
 		return nil, pkgHTTP.HumaProblem(err)
@@ -79,15 +76,15 @@ func (handler *MetadataIndexHandler) CreateMetadataIndexHuma(ctx context.Context
 		return nil, pkgHTTP.HumaProblem(err)
 	}
 
-	return &CreateMetadataIndexOutputHuma{Status: http.StatusCreated, Body: index}, nil
+	return &CreateMetadataIndexResponse{Status: http.StatusCreated, Body: index}, nil
 }
 
 // --- GET /settings/metadata-indexes (list) ------------------------------------
 
-// ListMetadataIndexesInputHuma advertises the entity_name query filter in the spec
+// ListMetadataIndexesRequest advertises the entity_name query filter in the spec
 // (doc-only, no validation tags) and captures the raw query via Resolve for the
 // shared core's http.ValidateParameters binder.
-type ListMetadataIndexesInputHuma struct {
+type ListMetadataIndexesRequest struct {
 	EntityName string `query:"entity_name" doc:"Optional entity name filter"`
 
 	// rawQuery is the request's parsed query, captured by Resolve. It is the binding
@@ -97,7 +94,7 @@ type ListMetadataIndexesInputHuma struct {
 
 // Resolve captures the raw query before the handler. It performs NO validation and
 // NEVER returns an error — canonical rejection stays in http.ValidateParameters.
-func (in *ListMetadataIndexesInputHuma) Resolve(ctx huma.Context) []error {
+func (in *ListMetadataIndexesRequest) Resolve(ctx huma.Context) []error {
 	u := ctx.URL()
 	in.rawQuery = u.Query()
 
@@ -107,7 +104,7 @@ func (in *ListMetadataIndexesInputHuma) Resolve(ctx huma.Context) []error {
 // queries rebuilds the map[string]string that http.ValidateParameters consumes,
 // matching Fiber's c.Queries() (last value wins for a repeated key, present-but-
 // empty keys included).
-func (in *ListMetadataIndexesInputHuma) queries() map[string]string {
+func (in *ListMetadataIndexesRequest) queries() map[string]string {
 	out := make(map[string]string, len(in.rawQuery))
 	for k, vs := range in.rawQuery {
 		if len(vs) == 0 {
@@ -121,47 +118,47 @@ func (in *ListMetadataIndexesInputHuma) queries() map[string]string {
 	return out
 }
 
-// ListMetadataIndexesOutputHuma carries the flat index slice verbatim (matching the
+// ListMetadataIndexesResponse carries the flat index slice verbatim (matching the
 // Fiber http.OK body — a JSON array, not a pagination envelope).
-type ListMetadataIndexesOutputHuma struct {
+type ListMetadataIndexesResponse struct {
 	Status int
 	Body   []*mmodel.MetadataIndex
 }
 
-// ListMetadataIndexesHuma delegates to the shared getAllMetadataIndexes core.
-func (handler *MetadataIndexHandler) ListMetadataIndexesHuma(ctx context.Context, in *ListMetadataIndexesInputHuma) (*ListMetadataIndexesOutputHuma, error) {
+// ListMetadataIndexes delegates to the shared getAllMetadataIndexes core.
+func (handler *MetadataIndexHandler) ListMetadataIndexes(ctx context.Context, in *ListMetadataIndexesRequest) (*ListMetadataIndexesResponse, error) {
 	indexes, err := handler.getAllMetadataIndexes(ctx, in.queries())
 	if err != nil {
 		return nil, pkgHTTP.HumaProblem(err)
 	}
 
-	return &ListMetadataIndexesOutputHuma{Status: http.StatusOK, Body: indexes}, nil
+	return &ListMetadataIndexesResponse{Status: http.StatusOK, Body: indexes}, nil
 }
 
 // --- DELETE /settings/metadata-indexes/entities/{entity_name}/key/{index_key} --
 
-// DeleteMetadataIndexInputHuma is the delete request envelope. Both path params are
+// DeleteMetadataIndexRequest is the delete request envelope. Both path params are
 // plain strings (no UUID, no format tag).
-type DeleteMetadataIndexInputHuma struct {
+type DeleteMetadataIndexRequest struct {
 	EntityName string `path:"entity_name" doc:"Entity name (organization, ledger, segment, account, portfolio, asset, account_type, transaction, operation, operation_route, transaction_route)"`
 	IndexKey   string `path:"index_key" doc:"Index key (metadata key, e.g. 'tier')"`
 }
 
-// DeleteMetadataIndexOutputHuma has NO Body field: paired with DefaultStatus 204 it
-// makes Huma emit a bodiless 204, matching the Fiber http.NoContent path.
-type DeleteMetadataIndexOutputHuma struct{}
+// DeleteMetadataIndexResponse has NO Body field: paired with DefaultStatus 204 it
+// makes Huma emit a bodiless 204.
+type DeleteMetadataIndexResponse struct{}
 
-// DeleteMetadataIndexHuma delegates to the shared deleteMetadataIndex core; returns
+// DeleteMetadataIndex delegates to the shared deleteMetadataIndex core; returns
 // a bodiless 204 on success.
-func (handler *MetadataIndexHandler) DeleteMetadataIndexHuma(ctx context.Context, in *DeleteMetadataIndexInputHuma) (*DeleteMetadataIndexOutputHuma, error) {
+func (handler *MetadataIndexHandler) DeleteMetadataIndex(ctx context.Context, in *DeleteMetadataIndexRequest) (*DeleteMetadataIndexResponse, error) {
 	if err := handler.deleteMetadataIndex(ctx, in.EntityName, in.IndexKey); err != nil {
 		return nil, pkgHTTP.HumaProblem(err)
 	}
 
-	return &DeleteMetadataIndexOutputHuma{}, nil
+	return &DeleteMetadataIndexResponse{}, nil
 }
 
-// RegisterMetadataIndexRoutes registers the three migrated metadata-index operations
+// RegisterMetadataIndexRoutes registers the three metadata-index operations
 // on the shared Huma API. The auth + tenant middleware chain for these routes is
 // attached at the Fiber level BEFORE the Huma terminal, not here.
 //
@@ -191,7 +188,7 @@ func RegisterMetadataIndexRoutes(api huma.API, h *MetadataIndexHandler, opSuffix
 		// Body validated imperatively (http.DecodeAndValidate) — see file header.
 		SkipValidateBody: true,
 		DefaultStatus:    http.StatusCreated,
-	}, h.CreateMetadataIndexHuma)
+	}, h.CreateMetadataIndex)
 	attachTypedRequestBody[mmodel.CreateMetadataIndexInput](api, "createMetadataIndex"+opSuffix)
 
 	huma.Register(api, huma.Operation{
@@ -201,7 +198,7 @@ func RegisterMetadataIndexRoutes(api huma.API, h *MetadataIndexHandler, opSuffix
 		Summary:     "Get all Metadata Indexes",
 		Tags:        []string{tag},
 		Security:    secMetadataBearerOrAPIKey,
-	}, h.ListMetadataIndexesHuma)
+	}, h.ListMetadataIndexes)
 
 	huma.Register(api, huma.Operation{
 		OperationID: "deleteMetadataIndex" + opSuffix,
@@ -212,10 +209,10 @@ func RegisterMetadataIndexRoutes(api huma.API, h *MetadataIndexHandler, opSuffix
 		Security:    secMetadataBearerOrAPIKey,
 		// DefaultStatus 204 + an Out struct with no Body field => bodiless 204.
 		DefaultStatus: http.StatusNoContent,
-	}, h.DeleteMetadataIndexHuma)
+	}, h.DeleteMetadataIndex)
 }
 
-// RegisterMetadataIndexRoutesToApp wires the Huma-migrated metadata-index surface onto the
+// RegisterMetadataIndexRoutesToApp wires the metadata-index surface onto the
 // /v1 contract. See registerMetadataIndexRoutesToApp for what it attaches.
 func RegisterMetadataIndexRoutesToApp(group fiber.Router, api huma.API, auth *middleware.AuthClient, h *MetadataIndexHandler, routeOptions *pkgHTTP.ProtectedRouteOptions) {
 	registerMetadataIndexRoutesToApp(group, api, auth, h, routeOptions, routeOpSuffixV1)
@@ -235,14 +232,13 @@ func RegisterMetadataIndexV2RoutesToApp(group fiber.Router, api huma.API, auth *
 // protectedMidaz(auth,"settings",verb) (= auth.Authorize("midaz","settings",verb) + tenant
 // PostAuthMiddlewares) — as MIDDLEWARE ONLY (no terminal) on the VERSIONED GROUP with
 // GROUP-RELATIVE paths, then registers the Huma terminals via RegisterMetadataIndexRoutes on
-// the SAME group's Huma API. This preserves the pre-Huma ("settings", verb) authz tuples —
-// the resource is "settings", NOT "metadata-indexes" — and tenant resolution BYTE-FOR-BYTE
-// on whichever version group it is mounted on; no metadata-index route becomes public.
+// the SAME group's Huma API. The authz tuples are ("settings", verb) — the resource is
+// "settings", NOT "metadata-indexes" — and tenant resolution runs on whichever version
+// group the surface is mounted on; no metadata-index route becomes public.
 //
-// No ParseUUIDPathParameters is attached: the pre-migration routes.go metadata-index ops
-// carried none (their path params — entity_name, index_key — are not UUIDs). The terminal
-// auth/tenant middleware calls c.Next(), advancing into the Huma terminal. The op order
-// (post, get, delete) matches routes.go.
+// No ParseUUIDPathParameters is attached: the path params — entity_name, index_key — are
+// not UUIDs. The terminal auth/tenant middleware calls c.Next(), advancing into the Huma
+// terminal.
 //
 // opSuffix distinguishes the operation IDs one version group publishes from another's —
 // see routeOpSuffixV1. Nothing else varies between contracts, so a change to the surface
