+++ components/ledger/internal/adapters/http/in/operation.go
@@ -9,7 +9,6 @@ import (
 
 	libObservability "github.com/LerianStudio/lib-observability/v2"
 	libOpentelemetry "github.com/LerianStudio/lib-observability/v2/tracing"
-	"github.com/gofiber/fiber/v3"
 	"github.com/google/uuid"
 	"go.mongodb.org/mongo-driver/v2/bson"
 
@@ -17,9 +16,9 @@ import (
 	"github.com/LerianStudio/midaz/v4/components/ledger/internal/services/command"
 	"github.com/LerianStudio/midaz/v4/components/ledger/internal/services/query"
 	"github.com/LerianStudio/midaz/v4/pkg/net/http"
-	// OperationHandler struct contains a cqrs use case for managing operations.
 )
 
+// OperationHandler holds the CQRS use cases that serve the operation resource.
 type OperationHandler struct {
 	Command *command.UseCase
 	Query   *query.UseCase
@@ -27,19 +26,15 @@ type OperationHandler struct {
 
 // --- Transport-agnostic cores -------------------------------------------------
 //
-// The two read cores below own the span, imperative query validation, the
-// metadata-vs-default branch, the service call and the pagination assembly. They
-// take primitive args (parsed UUIDs + the query map) so BOTH transports feed them:
-// the Fiber wrappers pull those from fiber.Ctx (Locals + c.Queries) and the Huma
-// handlers (operation_handler_huma.go) pull them from the request envelope. Every
-// canonical Midaz error the cores return is rendered by the caller — http.WithError
-// on the Fiber path, http.HumaProblem on the Huma path — so the code + HTTP status
-// are identical across both transports. Reads only; the command use case is
-// untouched.
+// The cores below own the span, imperative query validation, the
+// metadata-vs-default branch, the service calls and the pagination assembly. They
+// take primitive args (parsed UUIDs, the query map, the decoded payload); the Huma
+// handlers in operation_handler.go pull those from the request envelope. Every
+// canonical Midaz error a core returns is rendered by the caller through
+// http.HumaProblem, so the code + HTTP status stay canonical.
 
-// getAllOperationsByAccount binds the query imperatively (http.ValidateParameters —
-// the SAME binder the Fiber path used), preserves the metadata-vs-default branch,
-// then returns the cursor-paginated envelope.
+// getAllOperationsByAccount binds the query imperatively (http.ValidateParameters),
+// applies the metadata-vs-default branch, then returns the cursor-paginated envelope.
 func (handler *OperationHandler) getAllOperationsByAccount(ctx context.Context, organizationID, ledgerID, accountID uuid.UUID, queries map[string]string) (http.Pagination, error) {
 	_, tracer, _, _ := libObservability.NewTrackingFromContext(ctx)
 
@@ -108,104 +103,10 @@ func (handler *OperationHandler) getOperationByAccount(ctx context.Context, orga
 	return op, nil
 }
 
-// GetAllOperationsByAccount retrieves all operations by account.
-func (handler *OperationHandler) GetAllOperationsByAccount(c fiber.Ctx) error {
-	ctx := c.Context()
-
-	organizationID, err := http.GetUUIDFromLocals(c, "organization_id")
-	if err != nil {
-		return http.WithError(c, err)
-	}
-
-	ledgerID, err := http.GetUUIDFromLocals(c, "ledger_id")
-	if err != nil {
-		return http.WithError(c, err)
-	}
-
-	accountID, err := http.GetUUIDFromLocals(c, "account_id")
-	if err != nil {
-		return http.WithError(c, err)
-	}
-
-	pagination, err := handler.getAllOperationsByAccount(ctx, organizationID, ledgerID, accountID, c.Queries())
-	if err != nil {
-		return http.WithError(c, err)
-	}
-
-	return http.OK(c, pagination)
-}
-
-// GetOperationByAccount retrieves an operation by account.
-func (handler *OperationHandler) GetOperationByAccount(c fiber.Ctx) error {
-	ctx := c.Context()
-
-	organizationID, err := http.GetUUIDFromLocals(c, "organization_id")
-	if err != nil {
-		return http.WithError(c, err)
-	}
-
-	ledgerID, err := http.GetUUIDFromLocals(c, "ledger_id")
-	if err != nil {
-		return http.WithError(c, err)
-	}
-
-	accountID, err := http.GetUUIDFromLocals(c, "account_id")
-	if err != nil {
-		return http.WithError(c, err)
-	}
-
-	operationID, err := http.GetUUIDFromLocals(c, "operation_id")
-	if err != nil {
-		return http.WithError(c, err)
-	}
-
-	op, err := handler.getOperationByAccount(ctx, organizationID, ledgerID, accountID, operationID)
-	if err != nil {
-		return http.WithError(c, err)
-	}
-
-	return http.OK(c, op)
-}
-
-// UpdateOperation method that patch operation created before
-func (handler *OperationHandler) UpdateOperation(p any, c fiber.Ctx) error {
-	ctx := c.Context()
-
-	organizationID, err := http.GetUUIDFromLocals(c, "organization_id")
-	if err != nil {
-		return http.WithError(c, err)
-	}
-
-	ledgerID, err := http.GetUUIDFromLocals(c, "ledger_id")
-	if err != nil {
-		return http.WithError(c, err)
-	}
-
-	transactionID, err := http.GetUUIDFromLocals(c, "transaction_id")
-	if err != nil {
-		return http.WithError(c, err)
-	}
-
-	operationID, err := http.GetUUIDFromLocals(c, "operation_id")
-	if err != nil {
-		return http.WithError(c, err)
-	}
-
-	payload := p.(*operation.UpdateOperationInput)
-
-	op, err := handler.updateOperation(ctx, organizationID, ledgerID, transactionID, operationID, payload)
-	if err != nil {
-		return http.WithError(c, err)
-	}
-
-	return http.OK(c, op)
-}
-
 // updateOperation is the transport-neutral update core: it logs the safe payload,
-// runs command.UpdateOperation, then re-reads the operation via query.GetOperationByID
-// (mutable metadata/description only — amounts/accounts/direction/type are immutable).
-// Called by BOTH the Fiber wrapper and the Huma shell (operation_handler_huma.go). The
-// command use case is untouched (transport-only extraction).
+// runs command.UpdateOperation, then re-reads the operation via query.GetOperationByID.
+// Only metadata and description are mutable — amounts, accounts, direction and type
+// are immutable.
 func (handler *OperationHandler) updateOperation(ctx context.Context, organizationID, ledgerID, transactionID, operationID uuid.UUID, payload *operation.UpdateOperationInput) (*operation.Operation, error) {
 	logger, tracer, _, _ := libObservability.NewTrackingFromContext(ctx)
